+++ packages/functions/src/functions/auth/onDeleteUser.ts
@@ -1,5 +1,5 @@
 import { admin } from '../../modules/firebase';
-import { setUserToNotActive } from '../../utils/firestore/users';
+import { setUserToNotActive } from '../../utils/firestore/users/setUserToNotActive';
 
 export default async ({ uid }: admin.auth.UserRecord) => {
   await setUserToNotActive(uid);
