+++ packages/functions/src/utils/firestore/users/initializeUser.ts
@@ -1,5 +1,5 @@
 import { FirestoreDateLike, UserData } from '@yukukuru/types';
-import { firestore, admin } from '../../modules/firebase';
+import { firestore, admin } from '../../../modules/firebase';
 
 const collection = firestore.collection('users');
 
@@ -20,21 +20,5 @@ export async function initializeUser(id: string, props: Pick<UserData, 'photoUrl
     pausedGetFollower: false,
     ...props,
   };
-  await collection.doc(id).set(data, { merge: true });
-}
-
-/**
- * ユーザーのactiveをfalseにする
- */
-export async function setUserToNotActive(id: string): Promise<void> {
-  const data: Pick<UserData, 'active'> = { active: false };
-  await collection.doc(id).set(data, { merge: true });
-}
-
-/**
- * ユーザーのinvalidを変更
- */
-export async function updateUserInvalid(id: string, invalid: boolean): Promise<void> {
-  const data: Pick<UserData, 'invalid'> = { invalid };
   await collection.doc(id).set(data, { merge: true });
 }
